refactor: break the hook into more readable hooks
