+++ packages/api/src/hooks/p2p/useUpdateAdvertiserPaymentMethods.ts
@@ -1,0 +1,35 @@
+import { useCallback } from 'react';
+import useInvalidateQuery from '../../useInvalidateQuery';
+import useAuthorize from '../useAuthorize';
+import useMutation from '../../useMutation';
+
+type TPayloads = NonNullable<
+    NonNullable<Parameters<ReturnType<typeof useMutation<'p2p_advertiser_payment_methods'>>['mutate']>[0]>['payload']
+>;
+type TUpdatePayload = NonNullable<TPayloads['update']>[0];
+
+/** A custom hook that sends a request to update an existing p2p advertiser payment method. */
+const useUpdateAdvertiserPaymentMethods = () => {
+    const invalidate = useInvalidateQuery();
+    const { isSuccess } = useAuthorize();
+    const { mutate, ...rest } = useMutation('p2p_advertiser_payment_methods', {
+        onSuccess: () => invalidate('p2p_advertiser_payment_methods'),
+    });
+
+    const update = useCallback(
+        (id: string, values: TUpdatePayload) => {
+            if (isSuccess) {
+                mutate({ payload: { update: { [id]: { ...values } } } });
+            }
+        },
+        [mutate, isSuccess]
+    );
+
+    return {
+        /** Sends a request to update an existing p2p advertiser payment method */
+        update,
+        ...rest,
+    };
+};
+
+export default useUpdateAdvertiserPaymentMethods;
